requirements.txt
flask
python-dotenv
requests
cachetools
logging